refactor(data): Drinksベースのデータを別のStoreに切り出した see #170
Drinksベースのテーブルのコンポーネント名が抽象的すぎたため、具体的な名前に変更した
