+++ pages/data/monthly.vue
@@ -6,7 +6,7 @@ import { storeToRefs } from 'pinia'
 import { useMonthlyStore } from '~/store/pages/data/monthly'
 
 const monthlyStore = useMonthlyStore()
-const { yearMonth, chartDataTitle, computeCalendarData, computeGraphData, computedTableData, computedChartData, computedChartOptions, computedPieChartOptions } = storeToRefs(monthlyStore)
+const { yearMonth, computeCalendarData, computeGraphData, computedChartOptions } = storeToRefs(monthlyStore)
 const { prevMonth, nextMonth, fetchDrinkCounters } = monthlyStore
 
 // カレンダー再描画用
@@ -52,14 +52,11 @@ watch(yearMonth, async () => {
       :data="computeGraphData"
       :options="computedChartOptions"
     />
-    <PagesDataGraphsPieChart
-      :data="computedChartData"
-      :options="computedPieChartOptions"
-    />
-    <PagesDataTable
-      :headers="chartDataTitle"
-      :table-data="computedTableData"
-    />
+
+    <PagesDataGraphsPieChart />
+
+    <PagesDataAggregationByDrinksTable />
+
     <PagesDataAggregationByDowTable />
   </div>
 </template>
