+++ store/pages/data/components/types/aggregationByDrinksTable.ts
@@ -1,0 +1,4 @@
+export interface AggregationByDrink {
+  drink_id: number
+  count: number
+}
